Add/remove favorites from artist/album/playlist views
